+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Version 3.1
 Adding documents for ingredients
 Migration script added
+Formula picture is now stored in DB
 
 ### Version 3.0
 Add multiple suppliers per ingredient